Adding classes for Abstract Factory
This Abstract Factory specifically allows for different wrapper software
to be utilised.
IShellCommands.as - the Interface for SSCommands.as and ZincCommands.as
to adhere to.
ShellAbsractCreator.as - the abstract class that declares the methods to
implement.
ShellConcreteCreator.as - extends the abstract ShellAbstractCreator.as
class and is the class to be instantiated.
SSCommands.as - a Factory class that implements SWF Studio
functionality.  This is invoked from ShellConcreteCreator.as.
SSCommands.as - a Factory class that implements SWF Studio
functionality.    This is invoked from ShellConcreteCreator.as.